+++ src/main/java/com/pvpindex/factions/bootstrap/CommandsBootstrapComponent.java
@@ -85,7 +85,8 @@ public final class CommandsBootstrapComponent extends AbstractBootstrapComponent
         commandRegistry.register(new CmdSetHome(factionSvc, context.infra().getTerritoryGuard()));
         commandRegistry.register(new CmdUnsetHome(factionSvc));
         commandRegistry.register(new CmdFly(factionSvc));
-        commandRegistry.register(new CmdRelation(factionSvc));
+        commandRegistry.register(new CmdRelation(factionSvc, context.infra().getEzCountdownNotifier(),
+            context.infra().getNotificationsConfig()));
         commandRegistry.register(new CmdWarp(factionSvc, warpSvc, context.infra().getTerritoryGuard()));
         commandRegistry.register(new CmdBank(factionSvc, economy));
         commandRegistry.register(new CmdPower(context.infra().getVaultEconomy(), cfg, repos));